+++ jest.config.js
@@ -24,6 +24,8 @@ module.exports = {
     '<rootDir>/src/data/api/mocks.ts',
     '<rootDir>/src/data/db/mocks.ts',
     '<rootDir>/src/core/logger.ts',
+    '<rootDir>/src/core/cache.ts',
+    '<rootDir>/src/core/context.ts',
     '<rootDir>/src/gql/schema.ts',
     '<rootDir>/src/gql/launch/schema.ts',
     '<rootDir>/src/gql/mission/schema.ts',
@@ -50,6 +52,8 @@ module.exports = {
     '^@core/(.*)$': '<rootDir>/src/core/$1',
     '^@utils': '<rootDir>/src/utils.ts',
     '^@logger': '<rootDir>/src/core/logger.ts',
+    '^@cache': '<rootDir>/src/core/cache.ts',
+    '^@context': '<rootDir>/src/core/context.ts',
     '^@gql/(.*)$': '<rootDir>/src/gql/$1',
     '^@launchResolver': '<rootDir>/src/gql/launch/launchResolver.ts',
     '^@missionResolver': '<rootDir>/src/gql/mission/missionResolver.ts',
